Delete expired opportunities from ES. Move data_sync_check to new controller
