+++ spec/controllers/application_controller_spec.rb
@@ -6,10 +6,6 @@ RSpec.describe ApplicationController, type: :controller do
       expect(get: '/export-opportunities/check').to route_to(controller: 'application', action: 'check')
     end
 
-    it 'data sync check endpoint is publically accessible' do
-      expect(get: '/export-opportunities/data_sync_check').to route_to(controller: 'application', action: 'data_sync_check')
-    end
-
     it 'api check endpoint is publically accessible' do
       expect(get: '/export-opportunities/api_check').to route_to(controller: 'application', action: 'api_check')
     end
